refact: remove `constants` global variable
Befor this commit, the games "constants" was accessed through a global
variable (and used in both UI and server parts).

The problem with this design is that those values may not be updated if changed in
the data-base. (this is not currently possible but it is planned to be
implemented).

I have refactor so that those values are always read from the DB.

When those constants are used in "calculations" they are requested to
the DB with `getBreedingGameConstant()`

When those constants are shown in the UI, a "shiny module is used" to
ease the developement (cf. `./src/fun/constants_module.R`). All the
"servers outputs" for those values are defined in `src/server/server_constants.R`.