feat(admin): download and preview game initialisation report
This commit let the "admin" to see and download the game initaisation
report.

For that I also modify the behaviour of the app.
Now the `DATA_ROOT` folder **must** be present in order for the app to
start. The initialisation script will write in it this folder only if it
is empty (except for the `.gitkeep` file, and other hidden files that
are not checked).