feat(admin game init): add genetic values visualisation
For that I had to refactor how the genetic values are generated.
(cf. discussion on https://github.com/timflutre/PlantBreedGame/pull/37)

I used a simulation approach where I simulate the genotypes and
marker effects and then calculate the genetic values.

This is not the best in my opinion since the simulation of the genotypes
is a bit slow (~3s), fortunately the genotypes are simulated only once
(when the "game initialisation" windows apprears) as this simulation do
not depends an any parameter for now.

However, the marker effects simulation and genetic values calculation
are done every time `sigma_a2` or the pleiotropy parameters change.
This calculation takes a bit of time and the plots update is not 100%
smooth but it is acceptable (on my hardware).