optim(data-viz): avoid rendering plot 2 times when var type is changed
Previoulsy, `data_from_file` depended on both `input$categ_variables`
and `input$quant_variables`. This cause the plot to be calculated 2
times because when the user change a variable "type", eg. `Var_1` from
"categorical" to "quantitative" the following happens:
 - `data_from_file` is recalculated because `input$categ_variables`
 had changed
 - an observer update the value of `input$quant_variables`
 - `data_from_file` is again recalculated because `input$quant_variables`
 had changed

Now I have removed this dependency since all variables that are not
categorical are quantitative.

Note: the `input$quant_variables` input is not actually used anymore
and is here only as information for users.