Fix example request files
Previous file couldn't be uses "as is" with a fresh game setup because
the referenced individuals don't exist.
To make players easilly test the game, I modified those files to make
them use individuals for the initial collection. However, the specifyed
marker for single SNP genotyping may not always be available because
they are generated for each setups. I use one available with the default
parameters.